fase terminale

+++ ArbitralCoin_v1/app/Http/Controllers/PreferencesController.php
@@ -16,19 +16,25 @@ class PreferencesController extends Controller
         return view('tablePage.preferencesSettings')->with('logged',true)->with('loggedName',$_SESSION["loggedName"]);
     }
 
-    public function storeSettings(Request $request)
+    public function store(Request $request)
     {
         $dl = new DataLayer();
         $userID=$dl->getUserID($_SESSION["loggedEmail"]);
-        $dl->deleteFavExchanges($userID);
+
+        if($dl->findUserPreferencesByID($userID))
+        {
+            $dl->deleteFavExchanges($userID);
+            $dl->deleteUserPreferences($userID);
+        } 
+       
         
         $binance=isset($request['Binance']) ? $request['Binance'] : false;
         $kraken=isset($request['Kraken']) ? $request['Kraken'] : false;
         $crypto=isset($request['Crypto']) ? $request['Crypto'] : false;
-        $mockup=isset($request['Mockup']) ? $request['Mockup'] : false;
+        
         
         $dl->addUserPreferences($request->input('depositAmount'),$request->input('favoriteValute'),$request->input('minGain'),$userID);
-        $dl->addFavExchanges($binance,$kraken,$crypto,$mockup,$userID);
+        $dl->addFavExchanges($binance,$kraken,$crypto,$userID);
         return Redirect::to(route('bestPairs.index'));
     }
 }
